sinktest: Miscellaneous test rig improvements
This is a preperatory commit to use the workload checker to write into source
databases when building end-to-end integration tests.

It is now possible create a "swapped" base fixture, with the source and target
databases reversed. This base fixture can be further modified before
constructing an all.Fixture from it.

The workload generator uses 64-bit integer keys as a hedge against any future
key truncation.

Creating the test-table schema is broken up, allowing the table names and the
schema string expansion to happen in separate steps.

The generated payloads now include bodies for delete statements (used by
upcoming frontends) and a bug in the NewWebhookPayload() is fixed to elide the
data block for deletions.

+++ internal/util/workload/checker.go
@@ -28,8 +28,8 @@ import (
 // by a [GeneratorBase].
 type Checker struct {
 	*GeneratorBase
-	LoadChild    func(id int) (parent int, val int64, ok bool, err error)
-	LoadParent   func(id int) (val int64, ok bool, err error)
+	LoadChild    func(id int64) (parent int64, val int64, ok bool, err error)
+	LoadParent   func(id int64) (val int64, ok bool, err error)
 	RowCounter   func(table ident.Table) (int, error)
 	StageCounter func(table ident.Table, rng hlc.Range) (int, error)
 }
